requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
decorator==5.0.6
discord.py==1.6.0
disputils==0.1.3
dnspython==2.1.0
Flask==1.1.2
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jupyter-client==6.1.12
jupyter-core==4.7.1
MarkupSafe==1.1.1
motor==2.3.1
multidict==5.1.0
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
Pygments==2.8.1
pymongo==3.11.3
python-dateutil==2.8.1
pyzmq==22.0.3
replit==2.0.0
requests==2.25.1
schedule==0.6.0
six==1.15.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
yarl==1.6.3
